Rewrite to extract logic/configuration into a dedicated file
